# Conflicts:
#	Android/library/runtime/src/main/AndroidManifest.xml
#	CMakeLists.txt
#	Common/ac/gamesetupstruct.cpp
#	Common/ac/gamesetupstruct.h
#	Common/core/def_version.h
#	Common/game/interactions.cpp
#	Common/game/interactions.h
#	Compiler/script/cc_symboltable.h
#	Compiler/script/cs_compiler.cpp
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Editor/packages.config
#	Editor/AGS.Native/agsnative.cpp
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Editor/Native/ac/actiontype.cpp
#	Editor/Native/ac/actiontype.h
#	Editor/Native/ac/scriptmodule.cpp
#	Editor/Native/ac/scriptmodule.h
#	Engine/ac/event.cpp
#	Engine/ac/gamesetup.h
#	Engine/ac/gamestate.cpp
#	Engine/ac/gamestate.h
#	Engine/ac/global_display.cpp
#	Engine/ac/roomstatus.cpp
#	Engine/ac/roomstatus.h
#	Engine/ac/screen.cpp
#	Engine/game/savegame.cpp
#	Engine/game/savegame.h
#	Engine/game/savegame_components.h
#	Engine/gfx/ali3dogl.cpp
#	Engine/gfx/ali3dogl.h
#	Engine/main/config.cpp
#	Engine/main/engine.cpp
#	Engine/platform/windows/gfx/ali3dd3d.cpp
#	Solutions/AGS.Editor.Full.sln
#	Solutions/Engine.App/Engine.App.vcxproj.filters
#	iOS/xcode/ags/AGS.xcodeproj/project.pbxproj
#	version.json

+++ Compiler/fmem.cpp
@@ -56,8 +56,8 @@ void fmem_close(FMEM*cloo) {
 // of malloc and freeing
 #define BUFFER_INCREMENT_SIZE 2000
 
-void fmem_write(char* towr,long mei,FMEM*fill) {
-  long curdatlen=fill->len;
+void fmem_write(char* towr ,size_t mei, FMEM*fill) {
+  size_t curdatlen=fill->len;
   if (curdatlen + mei >= fill->size) {
     char*datwas=fill->data;
     fill->size=fill->size + mei + BUFFER_INCREMENT_SIZE;
@@ -79,7 +79,7 @@ void fmem_putc(char toput,FMEM*fill) {
   fmem_write(&templ[0],1,fill);
   }
 
-void fmem_puts(char*strin,FMEM*ooo) {
+void fmem_puts(char*strin, FMEM*ooo) {
   char*ibuffer=(char*)malloc(strlen(strin)+10);
   sprintf(ibuffer,"%s\r\n",strin);
   fmem_write(ibuffer,strlen(ibuffer),ooo);
